BXVGADevice+MBVGADevice: Correctly check page-aligned mmaps
In ab14b0ac64cd8bcaf7060050a7ec5a99cf7bd121, mmap was changed so that
the size of the region is aligned before it was passed to the device
driver. The previous logic would assert when the framebuffer size was
not a multiple of the page size. I've also taken the liberty of
returning an error on mmap failure rather than asserting.